plumbing: serverinfo, move to transport package
It makes more sense to have this living in the transport package, as it
is related to the transport layer.
